refactor: add reusable component for traveller review
